Add support for Berkshelf
This adds preferential support for Berkshelf instead of Librarian for
uploading a chef-repo to a chef-zero server.  If a Berksfile is present,
it will be used, otherwise Librarian is used to upload dependent
cookbooks.  Note that 'berks upload' is not used, as that may upload to
the real chef server, though this is a future optimization to consider.